refactor so can create base.GetDefaultTags
