+++ base/base.go
@@ -1,15 +1,12 @@
 package base
 
 import (
-	"bytes"
 	"context"
-	"errors"
-	"html/template"
 
-	"github.com/iter8-tools/etc3/api/v2alpha2"
+	"github.com/iter8-tools/handler/experiment"
+	"github.com/iter8-tools/handler/interpolation"
 	"github.com/iter8-tools/handler/utils"
 	"github.com/sirupsen/logrus"
-	corev1 "k8s.io/api/core/v1"
 )
 
 var log *logrus.Logger
@@ -44,90 +41,20 @@ func (a *Action) Run(ctx context.Context) error {
 	return nil
 }
 
-// Tags supports string extrapolation using tags.
-type Tags struct {
-	M map[string]interface{}
-}
-
-// NewTags creates an empty instance of Tags
-func NewTags() Tags {
-	return Tags{M: make(map[string]interface{})}
-}
-
-// WithSecret adds the fields in secret to tags
-func (tags Tags) WithSecret(secret *corev1.Secret) Tags {
-	if secret != nil {
-		for n, v := range secret.Data {
-			tags.M[n] = string(v)
+// GetDefaultTags creates interpolation.Tags from experiment referenced by context
+func GetDefaultTags(ctx context.Context) *interpolation.Tags {
+	tags := interpolation.NewTags()
+	exp, err := experiment.GetExperimentFromContext(ctx)
+	if err == nil {
+		obj, err := exp.ToMap()
+		if err == nil {
+			tags = tags.
+				With("this", obj).
+				WithRecommendedVersionForPromotion(&exp.Experiment)
 		}
-	}
-	return tags
-}
-
-// With adds obj to tags
-func (tags Tags) With(label string, obj interface{}) Tags {
-	if obj != nil {
-		tags.M[label] = obj
-	}
-	return tags
-}
-
-// WithRecommendedVersionForPromotion adds variables from versionDetail of version recommended for promotion
-func (tags Tags) WithRecommendedVersionForPromotion(exp *v2alpha2.Experiment) Tags {
-	if exp == nil || exp.Status.VersionRecommendedForPromotion == nil {
-		log.Warn("no version recommended for promotion")
-		return tags
-	}
-
-	versionRecommendedForPromotion := *exp.Status.VersionRecommendedForPromotion
-	if exp.Spec.VersionInfo == nil {
-		log.Warnf("No version details found for version recommended for promotion: %s", versionRecommendedForPromotion)
-		return tags
-	}
-
-	var versionDetail *v2alpha2.VersionDetail = nil
-	if exp.Spec.VersionInfo.Baseline.Name == versionRecommendedForPromotion {
-		versionDetail = &exp.Spec.VersionInfo.Baseline
 	} else {
-		for _, v := range exp.Spec.VersionInfo.Candidates {
-			if v.Name == versionRecommendedForPromotion {
-				versionDetail = &v
-				break
-			}
-		}
-	}
-	if versionDetail == nil {
-		log.Warnf("No version details found for version recommended for promotion: %s", versionRecommendedForPromotion)
-		return tags
+		log.Warn("No experiment found in context")
 	}
 
-	// get the variable values from the (recommended) versionDetail
-	tags.M["name"] = versionDetail.Name
-	for _, v := range versionDetail.Variables {
-		tags.M[v.Name] = v.Value
-	}
-
-	return tags
+	return &tags
 }
-
-// Interpolate str using tags.
-func (tags *Tags) Interpolate(str *string) (string, error) {
-	if tags == nil || tags.M == nil { // return a copy of the string
-		return *str, nil
-	}
-	var err error
-	var templ *template.Template
-	if templ, err = template.New("").Parse(*str); err == nil {
-		buf := bytes.Buffer{}
-		if err = templ.Execute(&buf, tags.M); err == nil {
-			return string(buf.Bytes()), nil
-		}
-		log.Error("template execution error: ", err)
-		return "", errors.New("cannot interpolate string")
-	}
-	log.Error("template creation error: ", err)
-	return "", errors.New("cannot interpolate string")
-}
-
-// ContextKey is the type of key that will be used to index into context.
-type ContextKey string
