+++ lib/metrics/metrics_env_test.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 
 	"github.com/iter8-tools/etc3/api/v2alpha2"
-	"github.com/iter8-tools/handler/base"
 	"github.com/iter8-tools/handler/experiment"
 	"github.com/iter8-tools/handler/utils"
 	. "github.com/onsi/ginkgo"
@@ -42,7 +41,7 @@ var _ = Describe("metrics library", func() {
 			}, exp2)).To(Succeed())
 
 			By("populating context with the experiment")
-			ctx := context.WithValue(context.Background(), base.ContextKey("experiment"), exp2)
+			ctx := context.WithValue(context.Background(), utils.ContextKey("experiment"), exp2)
 
 			By("creating a metrics/collect task")
 			ct := CollectTask{
@@ -117,7 +116,7 @@ var _ = Describe("metrics library", func() {
 			}, exp2)).To(Succeed())
 
 			By("populating context with the experiment")
-			ctx := context.WithValue(context.Background(), base.ContextKey("experiment"), exp2)
+			ctx := context.WithValue(context.Background(), utils.ContextKey("experiment"), exp2)
 
 			By("creating a metrics/collect task")
 			ct := CollectTask{
